[front] chore(eslint): Move single rule to its own file
Will make adding new rules easier
